Refactored to use configurable interface as optional
