+++ src/ConfigurableTamperBase.php
@@ -2,12 +2,62 @@
 
 namespace Drupal\tamper;
 
+use Drupal\Component\Utility\NestedArray;
 use Drupal\Core\Form\FormStateInterface;
 
 /**
  * Provides a configurable base class to tamper data from.
  */
 abstract class ConfigurableTamperBase extends TamperBase implements ConfigurableTamperInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function __construct(array $configuration, $plugin_id, $plugin_definition) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->setConfiguration($configuration);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getConfiguration() {
+    return $this->configuration;
+  }
+
+  /**
+   * Get a particular configuration value.
+   *
+   * @param string $key
+   *   Key of the configuration.
+   *
+   * @return mixed
+   *   Setting value.
+   */
+  public function getSetting($key) {
+    return $this->configuration[$key];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function setConfiguration(array $configuration) {
+    $this->configuration = NestedArray::mergeDeep($this->defaultConfiguration(), $configuration);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function defaultConfiguration() {
+    return [];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function calculateDependencies() {
+    return [];
+  }
 
   /**
    * {@inheritdoc}
